Moving DELETE /user into /user routes file
